page 합치기,  불필요한 페이지 삭제/수정
