+++ lib/util/pages.dart
@@ -4,7 +4,6 @@ import 'package:godroad/view/page/challenge_upload_page.dart';
 import 'package:godroad/view/page/login_detail_page.dart';
 import 'package:godroad/view/page/login_page.dart';
 import 'package:godroad/view/page/main_page.dart';
-import 'package:godroad/view/page/main_pageview_page.dart';
 import 'package:godroad/view/page/permission_page.dart';
 import 'package:godroad/view/page/edit_profile_page.dart';
 import 'package:godroad/view/page/my_page.dart';
@@ -18,7 +17,6 @@ class AppPages {
     GetPage(name: AppRoute.loginDetail, page: () => const LoginDetailPage()),
     GetPage(name: AppRoute.signup, page: () => const SignupPage()),
     GetPage(name: AppRoute.profile, page: () => const ProfilePage()),
-    GetPage(name: AppRoute.pageivew, page: () => const MainPageViewPage()),
     GetPage(name: AppRoute.permission, page: () => const PermissionPage()),
     GetPage(name: AppRoute.my, page: () => const MyPage()),
     GetPage(name: AppRoute.editProfile, page: () => const EditProfilePage()),
