+++ lib/util/routes.dart
@@ -2,7 +2,6 @@ import 'package:godroad/view/page/challenge_upload_page.dart';
 import 'package:godroad/view/page/login_detail_page.dart';
 import 'package:godroad/view/page/login_page.dart';
 import 'package:godroad/view/page/main_page.dart';
-import 'package:godroad/view/page/main_pageview_page.dart';
 import 'package:godroad/view/page/permission_page.dart';
 import 'package:godroad/view/page/edit_profile_page.dart';
 import 'package:godroad/view/page/my_page.dart';
@@ -15,7 +14,6 @@ class AppRoute {
   static final loginDetail = LoginDetailPage.route;
   static final signup = SignupPage.route;
   static final profile = ProfilePage.route;
-  static final pageivew = MainPageViewPage.route;
   static final permission = PermissionPage.route;
   static final my = MyPage.route;
   static final editProfile = EditProfilePage.route;
