+++ lib/view/page/outside_activity_screen.dart
@@ -1,0 +1,76 @@
+import 'package:flutter/material.dart';
+
+class OutsideActivityScreen extends StatelessWidget {
+  const OutsideActivityScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return ListView(
+      shrinkWrap: true,
+      children: [
+        Column(
+          mainAxisAlignment: MainAxisAlignment.start,
+          children: [
+            Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  const Text(
+                    '실시간 인기 대외 활동',
+                    style: TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                  TextButton(
+                    onPressed: () {},
+                    child: const Text('모두보기'),
+                  )
+                ],
+              ),
+            ),
+            SizedBox(
+              height: 200,
+              child: ListView(
+                shrinkWrap: true,
+                children: const [],
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: [
+                  const Text(
+                    '님을 위한 대외 활동',
+                    style: TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                  TextButton(
+                    onPressed: () {},
+                    child: const Text('모두보기'),
+                  )
+                ],
+              ),
+            ),
+            SizedBox(
+              height: 200,
+              child: ListView(
+                children: const [],
+              ),
+            ),
+            Padding(
+              padding: const EdgeInsets.all(16.0),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                children: const [
+                  Text(
+                    '키워드로 모아보기',
+                    style: TextStyle(fontWeight: FontWeight.bold),
+                  ),
+                ],
+              ),
+            )
+          ],
+        ),
+      ],
+    );
+  }
+}
